OAuth support
This adds the following method signatures to `ASIHTTPRequest` when
`ASIHTTPRequest+OAuth.h` is `#import`ed:

    + (id)requestWithURL:(NSURL *)newURL
             consumerKey:(NSString *)consumerKey
          consumerSecret:(NSString *)consumerSecret
                   token:(NSString *)token
             tokenSecret:(NSString *)tokenSecret;

    - (id)initWithURL:(NSURL *)newURL
          consumerKey:(NSString *)consumerKey
       consumerSecret:(NSString *)consumerSecret
                token:(NSString *)token
          tokenSecret:(NSString *)tokenSecret;

    - (void)addOAuthHeaderWithConsumerKey:(NSString *)consumerKey
                           consumerSecret:(NSString *)consumerSecret
                                    token:(NSString *)token
                              tokenSecret:(NSString *)tokenSecret
                          signatureMethod:(NSString *)signatureMethod;

    - (void)setOAuthSignatureMethod:(NSString *)signatureMethod;

When using `ASIFormDataRequest`, `#import "ASIFormDataRequest.h"` to enable
OAuth support for requests with bodies.

The patch does not require that bodies containing requests be created using
instances of `ASIFormDataRequest`; as long as a request body is present and the
content-type is specified as `application/x-www-form-urlencoded`, any
parameters contained within the body will be used to generate the signature.

Supported signature methods are `HMAC-SHA` and `PLAINTEXT`, with the former set
as the default.

To make 2-legged calls or requests for OAuth request tokens, use `nil` as the
value for `token` and `tokenSecret`.

To upgrade request tokens to access tokens, use the request token and secret as
the token values.

This includes an ASI-namespaced version of the Base64 transcoder from
[TouchFoundation](https://github.com/TouchCode/TouchFoundation) to avoid
conflicts with versions containing alternate method signatures that others have
incorporated into their applications.